Moved personal metrics to API endpoint
