+++ appinfo/routes.php
@@ -14,5 +14,6 @@ return [
 
 		['name' => 'subscription_change#list', 'url' => '/subscriptions', 'verb' => 'GET'],
 		['name' => 'subscription_change#create', 'url' => '/subscription_change/create', 'verb' => 'POST'],
+		['name' => 'personal_settings#metrics', 'url' => '/personal_settings/metrics', 'verb' => 'GET'],
 	]
 ];
